Prefer original file over symlink (v2)
If a symlink to a fontfile in the same directory exists the file
that got picked depends on the order in which readdir() picked
entries.
This patch gives the file a higher preference than the symlink to
it so it will be preferred if the xlfd entries match.

v2: Followed a suggestion by Julien Cristau <jcristau@debian.org>
    to check for DT_LNK being defined instead of _BSD_SOURCE
    and fall back to lstat() if d_type is DT_UNKNOWN.

Signed-off-by: Egbert Eich <eich@freedesktop.org>